Move handler.ts one level up
